Create twitter.php
I got so frustrated trying to complete the Twitter API section on the old version of The Complete Web Developer Course that I was finally able to work my way through the outdated material to find something that works. Probably could clean it up a bit, but felt it was more important to get something that works out there, in case there are still others working on it.